requirements.txt
# Core runtime dependencies
# NOTE: Keep this file de-duplicated to avoid confusing pip/resolution logs.

PyYAML>=6.0
python-dotenv>=1.0
numpy>=1.23
pandas>=2.0.0
pyarrow>=14.0.0
requests>=2.28
ccxt>=4.0.0

# Dev / test
pytest>=7.0

# UI (Control Panel)
customtkinter>=5.2.0
darkdetect>=0.8.0
psutil>=5.9.0
